move PagedResultTable story into ResultTable stories
